Restructures index.html and changes layout for css
index.html was changed to include div types for the recipe items in
recipes. The image for the recipe and the description and ingredient
list were made childs of a parent div. This was done so that the image
can be displayed to the left of the ingredient list and description.

css file was changed to incorporate the same changes for every recipe.
Each recipe has the has the same structure so the css file was made to
make the same layout for each recipe.

Also the main page was adjusted so that the title and all links to
recipes are centered. The list of recipes were changed to just regular
links to remove the dots from the list for better asthetics.